Migrate Violation to best practices
