+++ wemake_python_styleguide/violations/consistency.py
@@ -2520,46 +2520,3 @@ class RaiseSystemExitViolation(ASTViolation):
 
     error_template = 'Found `raise SystemExit`, instead of using `sys.exit`'
     code = 363
-
-
-@final
-class AwaitInLoopViolation(ASTViolation):
-    """
-    Forbid using ``await`` in ``for`` loop.
-
-    Reasoning:
-        There is a better way to control repeated coroutines in ``for`` loops.
-
-    Solution:
-        Using :function:`asyncio.gather` or :class:`asyncio.TaskGroup`
-        for Python 3.11+.
-
-    Example::
-
-        urls = [
-            "https://wemake-python-styleguide.readthedocs.io/",
-            "https://flake8.pycqa.org/",
-            "https://docs.astral.sh/ruff/",
-        ]
-
-        # Correct:
-        async def request():
-            tasks = []
-            for url in urls:
-                tasks.append(fetch_data(url))
-            results = await asyncio.gather(*tasks)
-
-        # Wrong:
-        async def request():
-            results = []
-            for url in urls:
-                result = await fetch_data(url)
-                results.append(result)
-        ...
-
-    .. versionadded:: 1.1.0
-
-    """
-
-    error_template = 'Found `await` in loop'
-    code = 364
