+++ wemake_python_styleguide/visitors/ast/loops.py
@@ -16,6 +16,7 @@ from wemake_python_styleguide.types import (
     AnyNodes,
 )
 from wemake_python_styleguide.violations.best_practices import (
+    AwaitInLoopViolation,
     InfiniteWhileLoopViolation,
     LambdaInsideLoopViolation,
     LoopVariableDefinitionViolation,
@@ -24,7 +25,6 @@ from wemake_python_styleguide.violations.complexity import (
     TooManyForsInComprehensionViolation,
 )
 from wemake_python_styleguide.violations.consistency import (
-    AwaitInLoopViolation,
     MultipleIfsInComprehensionViolation,
     UselessContinueViolation,
     WrongLoopIterTypeViolation,
